+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] - 2022-01-25
+
+### Fixed
+
+- Unchanged methods no longer have REACHABLE_BY edges regenerated/duplicated
+
 ## [1.0.4] - 2022-01-25
 
 ### Added